move ECS profiling serializer to ecsserializer package
